convert all file names to unix style in DZP::GatherDir
Ahead: All of the following assumes Win32 as the platform where these things were observed on.

The file names DZP::GD gathers were originally in platform-native format, this however meant two things:
1. the tests for the module failed because they expect unix-style paths
2. the MANIFEST was broken because it was written with the wrong windows style paths, including escaping and all

A better fix would probably be to write the test so it checks for platform-native paths and to make the MAKEFILE-writer convert to unix style when it writes. However, that is wasted effort in my opinion, since Windows understands unix-style paths just fine.