rename --inc (long form -I) to avoid conflict with --in
This is to cope with a bug/misconfig in Getopt::Long/Getopt::Long::Descriptive.

GLD allows the wrong (in my opinion) auto_abbrev option to be turned
on for Getopt::Long, which allows --in to be interpreted
by GL as an abbreviation for --inc.  I need to investigate whether
I can "just fix" GLD or need to make it an option.