PkgVersion and PkgDist no longer use BEGIN blocks
This behavior could lead to bugs on 5.8, where a module that failed to compile
would still appear in %INC, meaning subsequent C<require>s would not throw an
exception, and then checking ->VERSION would work because $VERSION was set.

Thanks, 5.8.  You're the best.

Meanwhile, this had been added to help with dynaloading code, which I have
since been told is not a big deal and not really a problem.  The BEGIN was just
left there to avoid unneeded changes.  Now the change is clearly needed.