Add 'exclude' option to GatherDir
This option excludes things from being gathered.  This is useful,
compared to PruneFiles, because gathering is early and pruning is late
and/or there might be a reason to not gather something that we don't
want pruned.

For example, if I want to keep a META.json copy in the root, I need to
exclude it from being gathered so that it isn't duplicated by MetaJSON.
PruneFiles doesn't help, because the gather is early and I do ultimately
want META.json in the distribution.