Use Test::Fatal in fakerelease.t
With Try::Tiny, if the code fails to die as expected, the catch block is never
executed and the test is just skipped (instead of failing as it should).  Since
there's no plan, the test script gets a PASS.  You're unlikely to notice that
it didn't run as many tests as it should.