authordeps: move uniq filter where the list is complete
The uniqueness filter is now applied to the plugin list once the full
list has been list instead of only after just the plugins from the
config sections have been listed. Previously if a plugin was also listed
both in a section name and with "; autordep" or in a @Filter it would
have been duplicated.