authordeps: move inc:: plugins at the beginning of the list
The order of the package list impacts how the loading of modules is
checked.
To give more control and more power to local inc:: plugins we move them
to the beginning of the list. That way, they will be able to tweak @INC.
The list of inc:: plugins is also sorted so the user can expect a
consistant order.