Anchor GatherDir's exclude_filename rules to avoid erroneous matches
The filenames specified for exclusion in GatherDir's exclude_filename
rule were added to the regex anchored with \b$filename\b, but that
is incorrect because filenames can have word boundaries. Since we're
matching against the whole path name, relative to the root, it
should suffice to anchor with ^ and $ instead.

Fixes RT#81492 for Dist-Zilla: GatherDir uses exclude_filename as
    a regex, causing files to be excluded incorrectly