Trim extraneous newline from version doc source.
Trim the newline off of the code fragment that's converted into the
PPI document that is the source of the Statement/Structure that is
inserted into the parent document.  The newline is unused (it ends up
in a different part of the PPI version document) and can be a bit
confusing when trying to figure out where the newlines in the output
document originate.

In other words, it's useless and slightly confusing.  Get rid of
it....