if a plugin sets dynamic_config to 1, it stays on
Any kind of dynamic configuration should make the whole dist
dynamically configured.  We don't want it turned off because some
other plugin claims not to be.  Maybe this is a mistake, and we
will later revert this so that a plugin can return 0 to override
a previous decision, but it seems highly unlikely.