PodCoverageTests test: unconditionnally load dependencies
This is a release test. Any missing release test dependency must be
handled as a failure of the test, not silently skipped.
And by the way, this case should not even happen as the develop/require
dependency is properly injected by the plugin, so the developer should
have installed them with 'dzil listdeps --author | cpanm'.