PruneCruft: prune fatlib/
Rationale: some author may build App-FatPaker-app from the project and
that created a fatlib/ directory that contains copies of CPAN modules.
If that fatlib/ is bundled in a distribution and indexed by PAUSE, that
will hide the original modules.
We want to avoid that.
See https://github.com/andk/pause/pull/65