PPI Role: prune end/data statements before trying to find variable assignments
The testcase I'm adding here would previously fail, and failed in the
wild on files that had POD followed by an __END__ marker if their
documentation showed an example of assigning to the $VERSION variable.

I.e. if you include POD after an __END__ marker PPI classifies the whole
POD as a statement.

I'm using dclone() here to clone the document, because the API is
somewhat sucky in that all the PPI documents are cached *and* used to
write the file out to disk, so if I munged it and didn't clone we'd
write a file with no POD or comments, which would suck.

I checked and we don't have to add a dependency on some Storable version
for this, we already depend on 5.8.5, and Storable already had dclone()
in 5.7.*.