Propagate exit code of dzil install
When the install command for dzil install fails an exception is
thrown. An eval captures this exception but does not rethrow it. Thus,
even an unsuccessful dzil install exits with code 0.

This patch logs the exception with log_fatal and thus dies in case of an
exception. This will set an exit code signaling an error.