Update critic-safeness changes to work with latest.
Things have changed since my last pull request.  The biggest change
was to make PkgVersion avoid adding extra lines, preferring to replace
an appropriately place blank line in the source.

This commit reworks the patch to respect this behaviour.

Lots of test cases and now they run twice, once with
die_on_line_insertion on and once with it off.