Utils/AuthorDeps.pm: re-write section splitting logic.
Authordeps had 2 largeish mistakes.

1. It figured it could strip everything after a trailing space in a
section name, and then use that stripped name as a key to extract data
from the original section hash and get the version.

This was however wrong, as after stripping the trailing space, it only
looked in an **empty** hash key, which of course, has no version
requirements.

ie:

    [Extra / Deps]
    :version = 1.0

Boiled down to

    $config->{"Extra"}->{':version'}

Which didn't exist.

So this resolves gh #353

2. It stripped on space, not on the slash delimiter.

This means it saw

    [Extra/Deps]

And parsed it as

    { pack => 'Extra/Deps', name => '' }

Where it should have parsed it as

    { pack => 'Extra', name => 'Deps' }

So this also resolves #196