DZ::Tester->from_config only accepts add_files entries with valid paths
Not all OSes have the same rules for what may be in a filename. By
round-tripping input from Unix to local to Unix and comparing the results
DZ::Tester now tries to determine whether a path is legal on the current
OS. This is not optimal, but works well enough to recognize some recently
introduced pathological cases.

Could maybe be made more strict by comparing whether the number of volume,
dirs, base entries in the objects is the same between all the participants.