add a PkgVersion test for double-package files under die_on_line_insertion
This is a regression test which ensures that a caching issue within PPI
cannot cause PkgVersion to fail to find the empty line after the second
package line in a file with two packages. This does not currently fail,
but until PPI (currently 1.218) is fixed to calculate locations correctly
after a tree was changed, this can easily break with seemingly logical
changes in DZ; making this test valuable.