don't slowly clone and prune in document_assigns_to_variable()
It is faster to skip END and DATA segments in this function than to clone
the document and prune it. I'm not sure if the order of the checks is
optimal. Putting the additional isa checks first might be faster. In any
case:

This provides a measurable speed-up on dzil build on dzil itself.