use release_status in place of is_trial
This change introduces a new attribute, 'release_status', and a new
role, D::Z::R::ReleaseStatusProvider.

The 'is_trial' variable is changed to be read-only.  This will break the
handful of CPAN modules that try to set it.  Those modules should set
the RELEASE_STATUS environment variable before building instead.

For backwards compatibility, 'is_trial' in dist.ini and the TRIAL
environment variable are supported.