Test::DZil: Load Dist::Zilla::Tester on demand
Although most people who load this module will be wanting the builder
functions, there are a couple of utility function in it ( namely,
simple_ini ) that are useful to have, even when not invoking main
Dzil code.

Subsequently, if you are testing an App::Command that doesn't normally
need to load dzil, but does some kind of processing on its own of
dist.ini ( Like authordeps does ), then you must presently pay the moose
price to do so.

This minor change avoids that penalty.

The one downside it does have is some people may have been inherently
dependending on DZT being loaded for them ( as was the test ), but that
is reasonably dodgy in itself.