use content, not encoded_content for PPI documents
I can't find the evidence for why we used encoded content, although I
was sure that there was some reason.  We found documents, though, that
can't be parsed when encoded, so we've switched to parsing the text
rather than the bytes until we have evidence of problems.

Thanks to Karen Etheridge for being unrelenting in making me revisit
this issue.