use quotes around eumm_version when needed
This check will fail when $ExtUtils::MakeMaker::VERSION is 6.55_02;

    use ExtUtils::MakeMaker 6.55_02;

..because older version.pm considered alpha versions less than their
numified equivalents. Consequently, we need to pass the version in as a
string, to ensure the proper comparison is made.

The proper thing to do would have been for the $VERSION declaration to
strip underscores, but some versions shipped without that -- e.g. the
version shipped with perl 5.10.1. :/