UploadToCPAN: only require a Login stash, not a PAUSE
PAUSE is really just a Login but you can put the credentials in the
configuration.  Why did I do this?  I have no idea, but I did.  If we
make it require a Login stash instead, we can use a stash with other
behavior, like fetching from an authenticated password manager.