UploadToCPAN: change how we report on problems with credentials
Previously, we'd report "no password provided" whether the password was
empty, never set, or even if the underlying password-getting mechanism
threw an exception.

We now distinguish between the two cases of "no password found" and "got
an exception".  Probably this can be further improved, but this fix
alone would've saved me a bunch of debugging time, so I'm sticking with
it.