drop use of Archive::Tar::Wrapper
We started to use Archive::Tar::Wrapper, when available, for speed.

It turns out that when the build directory contains non-ASCII
characters, this causes silent failures to build the tarball.  Rather
than diagnose and fix those, this removes Archive::Tar::Wrapper and
always uses the "slow" path.

The slow path is only 100ms slower, or so, in an already slow processes.
The cost of silent failures is not worth this speedup of some runs.